tests for foreign keys
