+++ library/src/test/java/com/github/gfx/android/orma/test/QueryTest.java
@@ -456,56 +456,6 @@ public class QueryTest {
     }
 
     @Test
-    public void updateSingleAssociation() throws Exception {
-        Publisher publisher = db.createPublisher(new ModelFactory<Publisher>() {
-            @NonNull
-            @Override
-            public Publisher call() {
-                Publisher publisher = new Publisher();
-                publisher.name = "The Nova";
-                publisher.startedYear = 2015;
-                publisher.startedMonth = 12;
-                return publisher;
-            }
-        });
-
-        int count = db.updateBook()
-                .where("title = ?", "today")
-                .publisher(publisher)
-                .execute();
-
-        assertThat(count, is(1));
-
-        Book book = db.selectFromBook().where("title = ?", "today").value();
-        assertThat(book.publisher.single().toBlocking().value().name, is("The Nova"));
-    }
-
-    @Test
-    public void updateSingleAssociationReference() throws Exception {
-        Publisher publisher = db.createPublisher(new ModelFactory<Publisher>() {
-            @NonNull
-            @Override
-            public Publisher call() {
-                Publisher publisher = new Publisher();
-                publisher.name = "The Nova";
-                publisher.startedYear = 2015;
-                publisher.startedMonth = 12;
-                return publisher;
-            }
-        });
-
-        int count = db.updateBook()
-                .where("title = ?", "today")
-                .publisher(SingleAssociation.<Publisher>id(publisher.id))
-                .execute();
-
-        assertThat(count, is(1));
-
-        Book book = db.selectFromBook().where("title = ?", "today").value();
-        assertThat(book.publisher.single().toBlocking().value().name, is("The Nova"));
-    }
-
-    @Test
     public void delete() throws Exception {
         int result = db.deleteFromBook()
                 .where("title = ?", "today")
@@ -799,65 +749,4 @@ public class QueryTest {
         assertThat(value, is(1L));
     }
 
-    @Test
-    public void testHasOne() throws Exception {
-        Publisher publisher = db.selectFromBook().value().publisher.single().toBlocking().value();
-        assertThat(publisher.name, is("foo bar"));
-        assertThat(publisher.startedYear, is(2015));
-        assertThat(publisher.startedMonth, is(12));
-    }
-
-    @Test
-    public void testHasManyRelation() throws Exception {
-        final Publisher a = db.createPublisher(new ModelFactory<Publisher>() {
-            @NonNull
-            @Override
-            public Publisher call() {
-                Publisher publisher = new Publisher();
-                publisher.name = "A";
-                return publisher;
-            }
-        });
-
-        final Publisher b = db.createPublisher(new ModelFactory<Publisher>() {
-            @NonNull
-            @Override
-            public Publisher call() {
-                Publisher publisher = new Publisher();
-                publisher.name = "B";
-                return publisher;
-            }
-        });
-
-        for (int i = 0; i < 2; i++) {
-            final int x = i;
-            db.createBook(new ModelFactory<Book>() {
-                @NonNull
-                @Override
-                public Book call() {
-                    Book book = new Book();
-                    book.publisher = SingleAssociation.id(a.id);
-                    book.title = "a " + x;
-                    return book;
-                }
-            });
-        }
-        for (int i = 0; i < 3; i++) {
-            final int x = i;
-
-            db.createBook(new ModelFactory<Book>() {
-                @NonNull
-                @Override
-                public Book call() {
-                    Book book = new Book();
-                    book.publisher = SingleAssociation.id(b.id);
-                    book.title = "b " + x;
-                    return book;
-                }
-            });
-        }
-
-        assertThat(a.books(db).count(), is(2));
-        assertThat(b.books(db).count(), is(3));
-    }
 }
